Move repository interfaces to SAPData.Repositories
