+++ SAPSec.Core/Features/Ks4CoreSubjects/UseCases/GetFilteredSchoolKs4CoreSubject.cs
@@ -1,8 +1,6 @@
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
-using SAPSec.Core.Features.SimilarSchools;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Data.Dto;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Features.Ks4CoreSubjects.UseCases;
