+++ SAPSec.Core/Features/Ks4CoreSubjects/UseCases/GetSchoolKs4CoreSubjects.cs
@@ -1,10 +1,8 @@
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
-using SAPSec.Core.Features.SimilarSchools;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Core.Model;
 using SAPSec.Data.Dto;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Features.Ks4CoreSubjects.UseCases;
