+++ SAPSec.Core/Features/Ks4HeadlineMeasures/UseCases/GetSchoolKs4HeadlineMeasures.cs
@@ -1,8 +1,7 @@
-using SAPSec.Core.Features.SimilarSchools;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Core.Model;
 using SAPSec.Data.Dto;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
