+++ SAPSec.Core/Features/SchoolSearch/SchoolSearchService.cs
@@ -1,9 +1,9 @@
 using SAPSec.Core.Constants;
 using SAPSec.Core.Features.Geography;
 using SAPSec.Core.Features.SchoolSearch.Extensions;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Data.Dto;
+using SAPSec.Data.Repositories;
 using System.Text.RegularExpressions;
 
 namespace SAPSec.Core.Features.SchoolSearch;
