+++ SAPSec.Core/Features/SimilarSchools/UseCases/FindSimilarSchools.cs
@@ -1,13 +1,11 @@
-using SAPSec.Core.Features.Attendance;
 using SAPSec.Core.Features.Filtering;
 using SAPSec.Core.Features.Geography;
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Core.Features.Pagination;
 using SAPSec.Core.Features.SimilarSchools.Filtering;
 using SAPSec.Core.Features.SimilarSchools.Sorting;
 using SAPSec.Core.Features.Sorting;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Model;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Features.SimilarSchools.UseCases;
