+++ SAPSec.Data/Repositories/IKs4PerformanceRepository.cs
@@ -1,6 +1,6 @@
 using SAPSec.Data.Dto.KS4.Performance;
 
-namespace SAPSec.Core.Features.Ks4HeadlineMeasures;
+namespace SAPSec.Data.Repositories;
 
 public interface IKs4PerformanceRepository
 {
