+++ SAPSec.Infrastructure/Json/JsonEstablishmentRepository.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Data.Dto;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Infrastructure.Json;
