+++ SAPSec.Infrastructure/Postgres/PostgresKs4PerformanceRepository.cs
@@ -1,7 +1,7 @@
 using Dapper;
 using Microsoft.Extensions.Logging;
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Data.Dto.KS4.Performance;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Infrastructure.Postgres;
