+++ Tests/SAPSec.Core.Tests/Features/Ks4CoreSubjects/GetSchoolKs4CoreSubjectsTests.cs
@@ -2,14 +2,12 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using SAPSec.Core.Features.Geography;
 using SAPSec.Core.Features.Ks4CoreSubjects.UseCases;
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
-using SAPSec.Core.Features.SimilarSchools;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Services;
 using SAPSec.Data.Dto;
 using SAPSec.Data.Dto.KS4.Performance;
 using SAPSec.Data.Dto.SimilarSchools.Secondary;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Tests.Features.Ks4CoreSubjects;
