+++ Tests/SAPSec.Core.Tests/Features/Ks4HeadlineMeasures/UseCases/GetKs4HeadlineMeasuresTests.cs
@@ -1,10 +1,10 @@
 using Moq;
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Core.Model;
 using SAPSec.Data.Dto.KS4.Destinations;
 using SAPSec.Data.Dto.KS4.Performance;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Tests.Features.Ks4HeadlineMeasures.UseCases;
