+++ Tests/SAPSec.Core.Tests/Features/Ks4HeadlineMeasures/UseCases/GetSchoolKs4HeadlineMeasuresTests.cs
@@ -1,14 +1,12 @@
 using Moq;
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
-using SAPSec.Core.Features.SimilarSchools;
-using SAPSec.Core.Interfaces.Repositories;
 using SAPSec.Core.Interfaces.Services;
 using SAPSec.Core.Model;
 using SAPSec.Data.Dto;
 using SAPSec.Data.Dto.KS4.Destinations;
 using SAPSec.Data.Dto.KS4.Performance;
 using SAPSec.Data.Dto.SimilarSchools.Secondary;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Core.Tests.Features.Ks4HeadlineMeasures.UseCases;
