+++ Tests/SAPSec.Test.Common/Repositories/InMemory/InMemoryAbsenceRepository.cs
@@ -1,5 +1,5 @@
-﻿using SAPSec.Core.Features.Attendance;
-using SAPSec.Data.Dto.Absence;
+﻿using SAPSec.Data.Dto.Absence;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Test.Common.Repositories.InMemory;
