+++ Tests/SAPSec.Test.Common/Repositories/InMemory/InMemorySimilarSchoolsSecondaryRepository.cs
@@ -1,5 +1,5 @@
-﻿using SAPSec.Core.Features.SimilarSchools;
-using SAPSec.Data.Dto.SimilarSchools.Secondary;
+﻿using SAPSec.Data.Dto.SimilarSchools.Secondary;
+using SAPSec.Data.Repositories;
 
 namespace SAPSec.Test.Common.Repositories.InMemory;
